Refactoring changes to prep for API v2.0

+++ Project2Api.Contracts/Order/OrderResponse.cs
@@ -7,6 +7,6 @@ namespace Project2Api.Contracts.Order
         Guid Id,
         DateTime dateTime,
         List<string> Items,
-        float Price
+        decimal Price
     );
 }
