+++ Project2Api.Contracts/Order/UpdateOrderRequest.cs
@@ -3,6 +3,6 @@ namespace Project2Api.Contracts.Order
     public record UpdateOrderRequest(
         DateTime dateTime,
         List<string> Items, 
-        float Price
+        decimal Price
     );
 }
